Working on images in ideas

+++ src/controllers/newsController.js
@@ -35,8 +35,15 @@ async function updateNews(req, res) {
 }
 exports.updateNews = updateNews;
 async function removeNews(req, res) {
-    let data = await newsService.removeNews(req.params.id);
-    return res.status(200).send(data);
+    const imageData = await newsService.getNewsById(req.params.id);
+    if (!imageData) {
+        return res.status(404).send();
+    }
+    if (imageData.avatarId) {
+        await imageService.removeImage(imageData.avatarId);
+    }
+    await newsService.removeNews(req.params.id);
+    return res.status(200).send();
 }
 exports.removeNews = removeNews;
 //# sourceMappingURL=newsController.js.map
